Create distancia reward sample
In this contract, the State struct maintains a list of adverts and rewards. The add_advert method is used to add an advert to the list, and the complete_advert method is used to mark an advert as completed. The reward_wallet method is used to reward the wallet address with a specified amount.

Note that this contract is only an example and may need to be modified and adapted to fit your specific use case.