Create EAKazi NFT Royalties-employer
This smart contract keeps track of NFTs using a HashMap that maps the ID of each NFT to the corresponding NFT struct. The NFT struct contains the wallet address of the issuer, as well as the amount of royalties earned by the issuer.

When the on_nft_used function is called, it increments the royalties for the issuer of the NFT by 2.5% of the total royalties earned by all NFT issuers. This ensures that the issuer of the NFT receives a share of the total royalties earned by all NFTs.