Create EAKazi NFT Royalties II
This contract defines a NFT struct that represents a non-fungible token, and a NFTContract struct that represents a contract for managing the NFTs. The NFTContract has a royalty_rate and an issuer_wallet that stores the issuer's wallet address.

The mint_nft function is used to mint a new NFT and add it to the contract. The transfer_nft function is used to transfer ownership of an NFT to a new owner. Finally, the use_nft function is used to reward the issuer's wallet address with a royalty payment whenever the NFT is used.

Note: This is just a simple example and is not meant to be used in a production environment. It may contain errors or omissions, and it is not tested or verified.